qualité : l'analyse statique repasse au vert, et ce qu'elle cachait
`golangci-lint` signalait onze problèmes. Il y en avait seize : le plafond par
défaut — trois occurrences d'un même message, cinquante par linter — masquait
les autres. La différence n'est pas anecdotique, c'est elle qui explique qu'un
défaut réel soit resté invisible.

Le défaut, justement. `SearchDescription` est le seul document XML de
l'OPDS construit à la main ; le flux, lui, passe par `encoding/xml`, qui échappe
de lui-même. Sans `baseURL` configurée, l'adresse interpolée dans
`template="…"` est composée à partir de l'en-tête `Host`, que le client choisit.
Un guillemet suffisait à sortir de l'attribut et à injecter du balisage dans un
document que des clients OPDS analysent. La valeur est désormais échappée.

Cinq `fmt.Errorf("%w : %v", ErrCorrupted, err)` dans les extracteurs d'archives
perdaient la seconde erreur : `errors.Is` ne retrouvait plus celle de `zip`,
`rar` ou `pdf`, et le diagnostic d'un fichier corrompu se réduisait à une
chaîne. Go 1.20 accepte plusieurs `%w`, forme déjà employée partout ailleurs
dans le dépôt.

Une comparaison `err != ErrTooLarge` dans les tests d'ingestion aurait cassé au
premier enveloppement de cette erreur — le genre de test qui devient faux sans
devenir rouge.

Quatre conversions `int → int32` non bornées. Deux sont bornées pour de bon :
la liste d'imports et la comparaison au catalogue local descendent toutes deux
d'un paramètre de requête, et leurs bornes vivaient chez l'appelant, donc par
convention. Elles sont maintenant dans le dépôt et dans la clôture, où elles
tiennent quel que soit l'appelant. Les deux autres — profondeur d'un dossier et
son écart — sont des invariants que l'analyseur ne peut pas voir : elles portent
un `#nosec` qui dit lequel, plutôt qu'une borne morte.

Le reste est du bruit assumé et documenté sur place : le retour de `fmt.Fprint`
sur l'aide en ligne, qu'aucun canal ne permettrait de signaler, et un « mot de
passe » qui est le nom d'un champ invalide dans un problème RFC 7807, pas un
secret.

+++ apps/server/internal/imaging/imaging_test.go
@@ -36,7 +36,7 @@ func planche(width, height int) image.Image {
 		shade := uint8(20 + i*18)
 		draw.Draw(img,
 			image.Rect(x, y, x+width/4, y+height/6),
-			&image.Uniform{color.RGBA{shade, uint8(255 - shade), shade, 255}},
+			&image.Uniform{color.RGBA{shade, 255 - shade, shade, 255}},
 			image.Point{}, draw.Src)
 	}
 	for y := 0; y < height; y++ {
